Version used in AMIA 2023 submission

+++ mhealth_anomaly_detection/plots.py
@@ -48,7 +48,7 @@ def lineplot_features(
         data.loc[f] = pd.to_numeric(data[f], errors="coerce")
         data.loc[time_col] = pd.to_numeric(data[time_col], errors="coerce")
         kwargs = {
-            "x": data.dropna(subset=[f])[time_col].astype(float),
+            "x": data.dropna(subset=[f])[time_col],
             "y": data.dropna(subset=[f])[f].astype(float),
             "color": color,
             "ax": ax,
